Refactor list_bundles methods
